clean up file strucutre in create
